getpitchPosition: use double query for MIDI pitch with diatonic interpolation

+++ include/vrv/toolkit.h
@@ -616,11 +616,11 @@ public:
      * Return the graphical coordinates of a pitch at a given score position and staff.
      *
      * @param scoreTime The score time (in quarter notes from start of the score)
-     * @param midiPitch MIDI pitch number
+     * @param midiPitch MIDI pitch number (fractional values allowed)
      * @param staffN The staff number (1-based)
      * @return A stringified JSON object with x / y coordinates and context information
      */
-    std::string GetPitchPosition(double scoreTime, int midiPitch, int staffN);
+    std::string GetPitchPosition(double scoreTime, double midiPitch, int staffN);
 
     ///@}
